backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.58.0
langgraph>=0.2.50
langchain>=0.3.0
pywebpush>=2.0.0
sentence-transformers>=3.3.0
apscheduler>=3.10.4
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
ruff>=0.8.0
mypy>=1.13.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
